% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Moments.R
\name{Moments}
\alias{Moments}
\title{SKR analysis - compute regression parameters of the SKR}
\usage{
Moments(
  values,
  individual_id = NULL,
  weight = NULL,
  values_factors = NULL,
  weight_factors = NULL,
  workers = parallel::detectCores() - 1
)
}
\arguments{
\item{values}{vector of values observed/measured per individual (length format)}

\item{individual_id}{vector of Individual name (length format, same length as values, must be completed if weight =! NULL and correspond to the column name in the weighting data frame)}

\item{weight}{dataframe of the weight to attribute to each individual for each factors of observation (e.g., frequency, abundance, ..., NO NA) (wide format, column name = Individual observed, must correspond to the row name in the values data frame)}

\item{values_factors}{dataframe of the factors of observations of the values vector}

\item{weight_factors}{dataframe of the factors of observations of the weight dataframe}

\item{workers}{Cores used for parallelisation (used 1 for no parallelisation)}
}
\value{
data.frame with:
mean of values per values_factors and/or weight_factors,
variance of values per values_factors and/or weight_factors,
skewness of values per values_factors and/or weight_factors,
kurtosis of values per values_factors and/or weight_factors,
}
\description{
Launch the SKR analysis of datasets with skewness and kurtosis values, based on explanatory factors (e.g., spatial, temporal, context, ...)
}
\examples{
head(values)
Moments(
values = values$Values,
individual_id = NULL,
weight = NULL,
values_factors = NULL,
weight_factors = NULL,
workers = 1
)
}
