% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SK.R
\docType{data}
\name{SK}
\alias{SK}
\title{Example dataset: SK}
\format{
A data frame with 9 rows and 6 variables:
\describe{
\item{Factor}{Principal Factor}
\item{Sub_Factor}{Sub Factor}
\item{mean}{mean}
\item{variance}{variance}
\item{skewness}{skewness}
\item{kurtosis}{kurtosis}
}
}
\source{
Generated internally
}
\usage{
SK
}
\description{
A dataset containing example moment values (mean, variance, skewness & kurtosis) used in the package according different Factors
}
\keyword{datasets}
