% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\docType{data}
\name{values}
\alias{values}
\title{Example dataset: values}
\format{
A data frame with 404 rows and 4 variables:
\describe{
\item{Factor}{Principal Factor}
\item{Sub_Factor}{Sub Factor}
\item{Individuals}{Statistic Individuals}
\item{Values}{Values observed/measured/simulated}
}
}
\source{
Generated internally
}
\usage{
values
}
\description{
A dataset containing example of Values per Individuals according different Factors
}
\keyword{datasets}
