% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SKRparam.R
\name{SKRparam}
\alias{SKRparam}
\title{SKR analysis - compute regression parameters of the SKR}
\usage{
SKRparam(
  moments,
  Factors = NULL,
  lin_mod = "lm",
  slope_ref = 1,
  intercept_ref = 1.86,
  distance_metric = "RMSE"
)
}
\arguments{
\item{moments}{dataframe of distribution moments with skewness and kurtosis columns}

\item{Factors}{dataframe of factors for the computation of statistics}

\item{lin_mod}{Indicates the type of linear model to use for (SKR): choose "lm" or "mblm"}

\item{slope_ref}{slope of a reference SKR used as a baseline (default: slope_ref = 1; skew-uniform slope)}

\item{intercept_ref}{intercept of a reference SKR used as a baseline (default: intercept_ref = 1.86; skew-uniform intercept)}

\item{distance_metric}{Indicates the method to compute distance-based regression parameters: choose "RMSE" (for Root Mean Square Error, default) or "MAE" (for Mean Absolute Error)}
}
\value{
data.frame with:
Slope per Factors,
Intercept per Factors,
Rsquare per Factors,
distance from predicted SKR per Factors,
CV of the distance from predicted SKR per Factors.
distance from reference SKR per Factors,
CV of the distance from reference SKR per Factors.
}
\description{
Launch the SKR analysis of datasets with skewness and kurtosis values, based on explanatory factors (e.g., spatial, temporal, context, ...)
}
\examples{
head(SK)
}
